Rename one of the versions to digital_crown_accessory_visibility
