--- Sources/LiveViewNative/Modifiers/Input Events Modifiers/DigitalCrownAccessoryModifier.swift
@@ -1,68 +1,0 @@
-//
-//  DigitalCrownAccessoryModifier.swift
-//  LiveViewNative
-//
-//  Created by Dylan.Ginsburg on 4/26/23.
-//
-
-import SwiftUI
-
-/// Specifies the visibility of Digital Crown accessory Views on Apple Watch.
-///
-/// ```html
-/// <Label modifiers={hover_effect(@native, effect: :lift)}>
-///     Do you lift bro
-/// </Label>
-/// ```
-///
-/// ## Arguments
-/// * ``effect``
-#if swift(>=5.8)
-@_documentation(visibility: public)
-#endif
-@available(watchOS 9.0, *)
-struct DigitalCrownAccessoryModifier: ViewModifier, Decodable {
-    /// The effect applied when the pointer hovers over a view.
-    ///
-    /// Possible values:
-    /// * `automatic`
-    /// * `highlight`
-    /// * `lift`
-    #if swift(>=5.8)
-    @_documentation(visibility: public)
-    #endif
-    private let effect: HoverEffect
-    
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-
-        #if os(iOS) || os(tvOS)
-        switch try container.decode(String.self, forKey: .effect) {
-        case "automatic": self.effect = .automatic
-        case "highlight": self.effect = .highlight
-        case "lift": self.effect = .lift
-        default:
-            throw DecodingError.dataCorruptedError(forKey: .effect, in: container, debugDescription: "invalid value for \(CodingKeys.effect.rawValue)")
-        }
-        #else
-        fatalError()
-        #endif
-    }
-    
-    func body(content: Content) -> some View {
-        content
-            #if os(iOS) || os(tvOS)
-            .hoverEffect(effect)
-            #endif
-    }
-    
-    enum CodingKeys: String, CodingKey {
-        case effect
-    }
-}
-
-extension HoverEffectModifier {
-    #if os(macOS) || os(watchOS)
-    typealias HoverEffect = Never
-    #endif
-}
