--- lib/live_view_native_swift_ui/modifiers/input_events/digital_crown_accessory.ex
@@ -1,7 +1,0 @@
-defmodule LiveViewNativeSwiftUi.Modifiers.DigitalCrownAccessory do
-  use LiveViewNativePlatform.Modifier
-
-  modifier_schema "digital_crown_accessory" do
-    field :visibility, Ecto.Enum, values: ~w(automatic visible hidden)a
-  end
-end
