+++ lib/live_view_native_swift_ui/modifiers/input_events/digital_crown_accessory_visibility.ex
@@ -1,0 +1,7 @@
+defmodule LiveViewNativeSwiftUi.Modifiers.DigitalCrownAccessoryVisibility do
+  use LiveViewNativePlatform.Modifier
+
+  modifier_schema "digital_crown_accessory_visibility" do
+    field :visibility, Ecto.Enum, values: ~w(automatic visible hidden)a
+  end
+end
